Separate timer functions from callbacks
